+++ packages/chip/src/lib/Chip.stories.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { Avatar, Collapse } from '@mui/material';
+import { Avatar } from '@availity/mui-avatar';
+import { Collapse } from '@mui/material';
 import { Chip, ChipProps } from './Chip';
 
 const meta: Meta<typeof Chip> = {
@@ -42,31 +43,4 @@ export const _RemovableChip: StoryObj<typeof Chip> = {
 
 export const _Avatar: StoryObj<typeof Chip> = {
   render: (args: ChipProps) => <Chip {...args} avatar={<Avatar>A</Avatar>} />,
-};
-
-/** Status styling is achieved by using size `small` */
-export const _Status: StoryObj<typeof Chip> = {
-  render: () => (
-    <>
-      Item 1 <Chip color="success" label="Approved" size="small" />
-      <br />
-      Item 2 <Chip color="info" label="Info" size="small" />
-      <br />
-      Item 3 <Chip color="warning" label="Pending" size="small" />
-      <br />
-      Item 4 <Chip color="error" label="Declined" size="small" />
-      <br />
-      Item 5 <Chip color="secondary" label="Neutral" size="small" />
-      <br />
-    </>
-  ),
-};
-
-/** Inline counter styling is achieved by using size `small`. For counters that affix to the corner of an element see `Badge`. */
-export const _InlineCounter: StoryObj<typeof Chip> = {
-  render: () => (
-    <div>
-      Items <Chip color="error" label="5" size="small" />
-    </div>
-  ),
 };