+++ packages/theme/src/lib/light-theme.ts
@@ -328,6 +328,10 @@ export const lightTheme = {
             color: tokens.colorPrimaryMain,
           },
         },
+        separator: {
+          marginLeft: '.625rem',
+          marginRight: '.625rem',
+        },
       },
     },
     MuiButton: {
@@ -342,15 +346,37 @@ export const lightTheme = {
           boxShadow: 'none',
           fontWeight: tokens.fontWeightsBold,
           textTransform: 'none',
+          lineHeight: '1.375rem',
+          fontSize: '.875rem',
         },
         containedPrimary: { ...containedButtonStyles('Primary') },
         containedSecondary: { ...containedButtonStyles('Secondary') },
         containedTertiary: {
-          ...containedButtonStyles('Tertiary'),
-          '&:focus': {
-            backgroundColor: tokens.colorTertiaryDark,
-            outline: '2px solid white',
-            boxShadow: `0 0 0px 4px ${tokens.colorSecondaryDark}`,
+          '&.MuiButton-containedTertiary': {
+            ...containedButtonStyles('Tertiary'),
+            boxSizing: 'border-box',
+            border: `1px solid ${tokens.borderSecondary}`,
+            color: tokens.colorTertiaryDark,
+            '&:hover': {
+              color: tokens.colorCommonWhite,
+              border: '1px solid transparent',
+              backgroundColor: tokens.colorTertiaryDark,
+            },
+            '&:active, &.Mui-focused:active, &:focus:active': {
+              color: tokens.colorCommonWhite,
+              border: '1px solid transparent',
+              backgroundColor: tokens.colorSecondaryLight,
+            },
+            '&.Mui-focused, &:focus': {
+              border: `1px solid ${tokens.borderSecondary}`,
+              backgroundColor: tokens.colorTertiaryLight,
+              color: tokens.colorTertiaryDark,
+            },
+            '&.Mui-disabled, &:disabled': {
+              border: `1px solid ${tokens.borderSecondary}`,
+              backgroundColor: tokens.colorTertiaryLight,
+              color: tokens.colorTertiaryDark,
+            },
           },
         },
         containedSuccess: { ...containedButtonStyles('Success') },
@@ -360,11 +386,54 @@ export const lightTheme = {
         outlinedSecondary: { ...outlinedButtonStyles('Secondary') },
         outlinedTertiary: {
           color: tokens.colorSecondaryMain,
-          ...outlinedButtonStyles('Secondary'),
-        }, // intended
+          '&.MuiButton-outlinedTertiary': {
+            ...containedButtonStyles('Tertiary'),
+            color: tokens.colorTertiaryDark,
+            '&:hover': {
+              color: tokens.colorCommonWhite,
+              border: '1px solid transparent',
+            },
+            '&:active, &.Mui-focused:active, &:focus:active': {
+              backgroundColor: tokens.colorSecondaryLight,
+              color: tokens.colorCommonWhite,
+              border: '1px solid transparent',
+            },
+            '&.Mui-focused, &:focus': {
+              border: `1px solid ${tokens.borderSecondary}`,
+              backgroundColor: tokens.colorTertiaryLight,
+              color: tokens.colorTertiaryDark,
+            },
+            '&.Mui-disabled, &:disabled': {
+              border: `1px solid ${tokens.borderSecondary}`,
+              backgroundColor: tokens.colorTertiaryLight,
+              color: tokens.colorTertiaryDark,
+            },
+          },
+        },
         outlinedSuccess: { ...outlinedButtonStyles('Success') },
         outlinedWarning: { ...outlinedButtonStyles('Warning') },
         outlinedError: { ...outlinedButtonStyles('Error') },
+        sizeSmall: {
+          padding: '4px 16px',
+          '&.MuiButton-containedTertiary, &.MuiButton-outlined': {
+            // any buttons with border need padding decreased to accommodate
+            padding: '3px 15px',
+          },
+        },
+        sizeMedium: {
+          padding: '6px 18px',
+          '&.MuiButton-containedTertiary, &.MuiButton-outlined': {
+            padding: '5px 17px',
+          },
+        },
+        sizeLarge: {
+          fontSize: '.9375rem',
+          lineHeight: '1.5rem',
+          padding: '8px 22px',
+          '&.MuiButton-containedTertiary, &.MuiButton-outlined': {
+            padding: '7px 21px',
+          },
+        },
       },
     },
     MuiCard: {
@@ -430,24 +499,6 @@ export const lightTheme = {
       styleOverrides: {
         root: {
           height: 'auto',
-          // Bug with state styles being overwritten when not defined in root https://github.com/mui/material-ui/issues/29703#issuecomment-1548356589
-          '&.MuiChip-deletable': {
-            '&:hover': {
-              backgroundColor: tokens.colorActionFocus,
-              boxShadow: 'none',
-              '.MuiChip-deleteIcon': {
-                color: tokens.colorSecondaryLight,
-              },
-            },
-            '&.Mui-focusVisible': {
-              backgroundColor: tokens.colorActionFocus,
-              outline: '2px solid white',
-              boxShadow: `0 0 0px 4px ${tokens.colorSecondaryLight}`,
-              '.MuiChip-deleteIcon': {
-                color: tokens.colorSecondaryLight,
-              },
-            },
-          },
         },
         avatar: {
           margin: '0',
@@ -460,24 +511,58 @@ export const lightTheme = {
         deleteIcon: {
           fontSize: '1rem',
           margin: '0',
-          transition: 'color 300ms',
+          fill: tokens.colorGrey300,
+          '&:hover': {
+            fill: tokens.colorGrey700,
+          },
         },
         label: {
           fontSize: 'inherit',
           padding: '0px 6px',
         },
+        labelSmall: {
+          padding: '0px 8px',
+          fontSize: '.75rem',
+          lineHeight: '1.125rem',
+        },
         colorDefault: {
-          backgroundColor: tokens.colorGrey100,
+          backgroundColor: tokens.colorBackgroundAccent,
+          '&:focus': {
+            backgroundColor: tokens.colorActionFocus,
+          },
         },
         sizeSmall: {
-          fontWeight: tokens.fontWeightsBold,
-          fontSize: '.75rem',
           verticalAlign: 'text-bottom',
           borderRadius: '50px',
         },
         sizeMedium: {
           fontWeight: tokens.fontWeightsRegular,
           padding: '3px 4px',
+        },
+      },
+    },
+    MuiPickersCalendarHeader: {
+      styleOverrides: {
+        root: {
+          padding: '4px 16px 8px',
+        },
+      },
+    },
+    MuiDayPicker: {
+      styleOverrides: {
+        monthContainer: {
+          padding: '12px 8px 0px',
+        },
+      },
+    },
+    MuiCalendarPicker: {
+      styleOverrides: {
+        root: {
+          height: '328px',
+          width: '310px',
+          '.MuiIconButton-root': {
+            border: '0',
+          },
         },
       },
     },
@@ -544,6 +629,7 @@ export const lightTheme = {
           display: important('flex'),
           flexDirection: important('row-reverse'),
           justifyContent: important('flex-end'),
+          marginBottom: '3px',
           '&.Mui-disabled': {
             color: tokens.colorTextDisabled,
           },
@@ -618,17 +704,41 @@ export const lightTheme = {
             },
             '&.Mui-disabled > fieldset': {
               borderColor: tokens.borderDisabled,
+              color: tokens.colorTextDisabled,
+            },
+            '.MuiInputBase-input': {
+              padding: '8px 12px',
+            },
+            '&.MuiInputBase-adornedEnd': {
+              paddingRight: '12px',
+            },
+            '&:hover': {
+              '.MuiIconButton-root': {
+                color: tokens.colorGrey700,
+              },
+            },
+            '&.Mui-disabled': {
+              '.MuiIconButton-root': {
+                color: tokens.colorTextDisabled,
+              },
+            },
+            '.MuiIconButton-root': {
+              border: '0',
+              padding: '8px',
+              color: tokens.colorGrey400,
             },
           },
-          '&.MuiButtonBase-root': {
-            padding: '8px 8px',
-          },
         },
-        input: {
-          width: '196px',
-          height: '24px',
-          borderRadius: '4px',
-          color: tokens.colorGrey400,
+        input: ({ ownerState }: { ownerState: { size: 'small' | 'medium' } }) => {
+          return {
+            width: '196px',
+            height: ownerState.size === 'small' ? '24px' : '40px',
+            borderRadius: '4px',
+            color: tokens.colorGrey400,
+            '&:hover': {
+              color: tokens.colorGrey700,
+            },
+          };
         },
         inputSizeSmall: {
           width: '100%',
@@ -652,9 +762,6 @@ export const lightTheme = {
           color: tokens.colorTextLink,
           '&:active': {
             color: tokens.colorTextPrimary,
-          },
-          '&:visited': {
-            color: tokens.colorIndigo600,
           },
         },
       },