+++ packages/tabs/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.6](https://github.com/Availity/element/compare/@availity/mui-tabs@0.1.5...@availity/mui-tabs@0.1.6) (2023-11-08)
+
+### Dependency Updates
+
+- `mui-divider` updated to version `0.3.4`
+
+## [0.1.5](https://github.com/Availity/element/compare/@availity/mui-tabs@0.1.4...@availity/mui-tabs@0.1.5) (2023-11-07)
+
 ## [0.1.4](https://github.com/Availity/element/compare/@availity/mui-tabs@0.1.3...@availity/mui-tabs@0.1.4) (2023-10-06)
 
 ## [0.1.3](https://github.com/Availity/element/compare/@availity/mui-tabs@0.1.2...@availity/mui-tabs@0.1.3) (2023-08-23)