+++ packages/textfield/CHANGELOG.md
@@ -2,12 +2,40 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## [0.2.0](https://github.com/Availity/element/compare/@availity/mui-textfield@0.1.1...@availity/mui-textfield@0.2.0) (2023-10-20)
+## [0.3.2](https://github.com/Availity/element/compare/@availity/mui-textfield@0.3.1...@availity/mui-textfield@0.3.2) (2023-11-08)
 
+### Dependency Updates
+
+- `mui-form-utils` updated to version `0.5.3`
+- `mui-button` updated to version `0.5.0`
+
+## [0.3.1](https://github.com/Availity/element/compare/@availity/mui-textfield@0.3.0...@availity/mui-textfield@0.3.1) (2023-11-08)
+
+### Dependency Updates
+
+- `mui-form-utils` updated to version `0.5.2`
+- `mui-icon` updated to version `0.7.2`
+- `mui-button` updated to version `0.4.5`
+
+## [0.3.0](https://github.com/Availity/element/compare/@availity/mui-textfield@0.2.1...@availity/mui-textfield@0.3.0) (2023-11-08)
 
 ### Features
 
-* **mui-textfield:** add select, wrap mui textfield ([c3329f9](https://github.com/Availity/element/commit/c3329f9ab6de35be3c07168f1d609c41fa29f79a))
+- **mui-datepicker:** add datepicker component ([2efdbf5](https://github.com/Availity/element/commit/2efdbf5112583c4322d588823f0ce0b7227fde6f))
+- **mui-datepicker:** update styles for datepicker ([b250d38](https://github.com/Availity/element/commit/b250d382691a83bfc12c066a2d4eeb9af147689f))
+
+### Bug Fixes
+
+- **mui-datepicker:** update styles ([1c0aa8a](https://github.com/Availity/element/commit/1c0aa8a3a2ce13df6655561236ebfaacdaceb28e))
+- **mui-datepicker:** use updated TextField ([23caa3a](https://github.com/Availity/element/commit/23caa3a0852f0249afb9ca0ef411db106c7049df))
+
+## [0.2.1](https://github.com/Availity/element/compare/@availity/mui-textfield@0.2.0...@availity/mui-textfield@0.2.1) (2023-11-07)
+
+## [0.2.0](https://github.com/Availity/element/compare/@availity/mui-textfield@0.1.1...@availity/mui-textfield@0.2.0) (2023-10-20)
+
+### Features
+
+- **mui-textfield:** add select, wrap mui textfield ([c3329f9](https://github.com/Availity/element/commit/c3329f9ab6de35be3c07168f1d609c41fa29f79a))
 
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-textfield@0.1.0...@availity/mui-textfield@0.1.1) (2023-10-06)