+++ packages/theme/CHANGELOG.md
@@ -2,26 +2,64 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## [0.22.0](https://github.com/Availity/element/compare/@availity/theme@0.21.0...@availity/theme@0.22.0) (2023-10-20)
+## [0.25.0](https://github.com/Availity/element/compare/@availity/theme@0.24.1...@availity/theme@0.25.0) (2023-11-08)
 
+### Dependency Updates
+
+- `design-tokens` updated to version `0.12.0`
 
 ### Features
 
-* **theme:** add legacy button styles ([9729a81](https://github.com/Availity/element/commit/9729a81116447f32c944d1b79c6daf617bd54a09))
+- **theme:** update tertiary button styles ([242982a](https://github.com/Availity/element/commit/242982a9a853e6646281f2becf217c605f4a5213))
+
+## [0.24.1](https://github.com/Availity/element/compare/@availity/theme@0.24.0...@availity/theme@0.24.1) (2023-11-08)
+
+### Dependency Updates
+
+- `design-tokens` updated to version `0.11.1`
+
+## [0.24.0](https://github.com/Availity/element/compare/@availity/theme@0.23.0...@availity/theme@0.24.0) (2023-11-08)
+
+### Features
+
+- **mui-datepicker:** add datepicker component ([2efdbf5](https://github.com/Availity/element/commit/2efdbf5112583c4322d588823f0ce0b7227fde6f))
+- **mui-datepicker:** update styles for datepicker ([b250d38](https://github.com/Availity/element/commit/b250d382691a83bfc12c066a2d4eeb9af147689f))
+
+### Bug Fixes
+
+- **mui-datepicker:** update padding and add help topic ([efd3487](https://github.com/Availity/element/commit/efd3487fbc6c01b9fe48187da238b8137fde1c2e))
+- **mui-datepicker:** update styles ([1c0aa8a](https://github.com/Availity/element/commit/1c0aa8a3a2ce13df6655561236ebfaacdaceb28e))
+- **mui-datepicker:** use updated TextField ([23caa3a](https://github.com/Availity/element/commit/23caa3a0852f0249afb9ca0ef411db106c7049df))
+
+## [0.23.0](https://github.com/Availity/element/compare/@availity/theme@0.22.1...@availity/theme@0.23.0) (2023-11-02)
+
+### Features
+
+- **mui-chip:** break out StatusChip into a separate component ([a1ac4db](https://github.com/Availity/element/commit/a1ac4dba9c944ae7ebe3cbbbd9b1b97a89abeb11))
+
+## [0.22.1](https://github.com/Availity/element/compare/@availity/theme@0.22.0...@availity/theme@0.22.1) (2023-10-25)
+
+### Bug Fixes
+
+- **mui-breadcrumbs:** home link target and separator tweaks ([4eb1b5f](https://github.com/Availity/element/commit/4eb1b5f172d072fe14ac9dc5b9d22707de7ca0b5))
+
+## [0.22.0](https://github.com/Availity/element/compare/@availity/theme@0.21.0...@availity/theme@0.22.0) (2023-10-20)
+
+### Features
+
+- **theme:** add legacy button styles ([9729a81](https://github.com/Availity/element/commit/9729a81116447f32c944d1b79c6daf617bd54a09))
 
 ## [0.21.0](https://github.com/Availity/element/compare/@availity/theme@0.20.0...@availity/theme@0.21.0) (2023-10-20)
 
-
 ### Features
 
-* **mui-form-utils:** add FormHelperText and Select ([0ff9d8e](https://github.com/Availity/element/commit/0ff9d8e06d0baac263ce1869304d3b91fcb5cb15))
+- **mui-form-utils:** add FormHelperText and Select ([0ff9d8e](https://github.com/Availity/element/commit/0ff9d8e06d0baac263ce1869304d3b91fcb5cb15))
 
 ## [0.20.0](https://github.com/Availity/element/compare/@availity/theme@0.19.4...@availity/theme@0.20.0) (2023-10-17)
 
-
 ### Features
 
-* **mui-card:** add legacy theme ([ff7e238](https://github.com/Availity/element/commit/ff7e2381c686f27cfc6ada02b1a4c315a28c90a5))
+- **mui-card:** add legacy theme ([ff7e238](https://github.com/Availity/element/commit/ff7e2381c686f27cfc6ada02b1a4c315a28c90a5))
 
 ## [0.19.4](https://github.com/Availity/element/compare/@availity/theme@0.19.3...@availity/theme@0.19.4) (2023-10-17)