+++ packages/typography/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.4](https://github.com/Availity/element/compare/@availity/mui-typography@0.1.3...@availity/mui-typography@0.1.4) (2023-11-08)
+
+### Dependency Updates
+
+- `mui-layout` updated to version `0.1.2`
+
+## [0.1.3](https://github.com/Availity/element/compare/@availity/mui-typography@0.1.2...@availity/mui-typography@0.1.3) (2023-11-07)
+
 ## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-typography@0.1.1...@availity/mui-typography@0.1.2) (2023-10-06)
 
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-typography@0.1.0...@availity/mui-typography@0.1.1) (2023-10-03)